Adds a freezing pass that will constant fold parameters in inductor `config.freezing`. This occurs post functionalization in aot autograd to capture both dispatching and allow passes to occur post functionalization. A few notes:

- There is an option to discard parameters `config.freezing_discard_parameters` which will take the current eager modules and wrap parameters to a Tensor subclass which will error if used. 
- I needed to expose flat_params in aot_autograd in order to discard old references when we constant fold away parameters, like with amp. I also exposed `fw_metadata` to avoid constant folding mutated paraemters. 
- Caching parameter transformations/constant folding across different inferences nyi
- Checking version_counter of constant folded params nyi


I'm not really sure what the actual naming should be. In jit there was both "freezing", which was platform agnostic, and "optimize for inference", which made device specific optimizations. We're doing the latter here but maybe freezing is a better name.

cc soumith voznesenskym penguinwu EikanWang jgong5 Guobing-Chen XiaobingSuper zhuhaozhe blzheng Xia-Weiwen wenzhe-nrv jiayisunx peterbell10 desertfire

Differential Revision: [D46244033](https://our.internmc.facebook.com/intern/diff/D46244033)

[ghstack-poisoned]

+++ test/run_test.py
@@ -29,6 +29,7 @@ from torch.testing._internal.common_utils import (
     retry_shell,
     set_cwd,
     shell,
+    TEST_WITH_ASAN,
     TEST_WITH_ROCM,
 )
 from torch.utils import cpp_extension
@@ -43,6 +44,7 @@ try:
         calculate_shards,
         get_reordered_tests,
         get_test_case_configs,
+        log_time_savings,
         NUM_PROCS,
         ShardedTest,
         THRESHOLD,
@@ -59,6 +61,7 @@ except ImportError:
 RERUN_DISABLED_TESTS = os.getenv("PYTORCH_TEST_RERUN_DISABLED_TESTS", "0") == "1"
 CPP_TEST_PREFIX = "cpp"
 CPP_TEST_PATH = "build/bin"
+DISTRIBUTED_TEST_PREFIX = "distributed"
 
 
 # Note [ROCm parallel CI testing]
@@ -424,7 +427,7 @@ JIT_EXECUTOR_TESTS = [
     "test_jit_fuser_legacy",
 ]
 
-DISTRIBUTED_TESTS = [test for test in TESTS if test.startswith("distributed")]
+DISTRIBUTED_TESTS = [test for test in TESTS if test.startswith(DISTRIBUTED_TEST_PREFIX)]
 FUNCTORCH_TESTS = [test for test in TESTS if test.startswith("functorch")]
 ONNX_TESTS = [test for test in TESTS if test.startswith("onnx")]
 CPP_TESTS = [test for test in TESTS if test.startswith(CPP_TEST_PREFIX)]
@@ -491,6 +494,7 @@ def run_test(
         test_file = test_module.name
         use_sharded_test = True
 
+    is_distributed_test = test_file.startswith(DISTRIBUTED_TEST_PREFIX)
     is_cpp_test = test_file.startswith(CPP_TEST_PREFIX)
     # NB: Rerun disabled tests depends on pytest-flakefinder and it doesn't work with
     # pytest-cpp atm. We also don't have support to disable C++ test yet, so it's ok
@@ -528,7 +532,12 @@ def run_test(
     # If using pytest, replace -f with equivalent -x
     if options.pytest:
         unittest_args.extend(
-            get_pytest_args(options, stepcurrent_key, is_cpp_test=is_cpp_test)
+            get_pytest_args(
+                options,
+                stepcurrent_key,
+                is_cpp_test=is_cpp_test,
+                is_distributed_test=is_distributed_test,
+            )
         )
         unittest_args = [arg if arg != "-f" else "-x" for arg in unittest_args]
 
@@ -586,7 +595,11 @@ def run_test(
     os.close(log_fd)
 
     command = (launcher_cmd or []) + executable + argv
-    should_file_rerun = "--subprocess" not in command and not RERUN_DISABLED_TESTS
+    should_file_rerun = (
+        "--subprocess" not in command
+        and not RERUN_DISABLED_TESTS
+        and not options.continue_through_error
+    )
     timeout = (
         THRESHOLD * 3
         if should_file_rerun
@@ -912,7 +925,7 @@ def print_log_file(test: str, file_path: str, failed: bool) -> None:
     num_lines = sum(1 for _ in open(file_path, "rb"))
     test = str(test)
     n = 100
-    with open(file_path, "r") as f:
+    with open(file_path, "rb") as f:
         print_to_stderr("")
         if failed:
             if n < num_lines:
@@ -923,25 +936,32 @@ def print_log_file(test: str, file_path: str, failed: bool) -> None:
                     f"##[group]PRINTING BEGINNING OF LOG FILE of {test} ({file_path})"
                 )
                 for _ in range(num_lines - n):
-                    print_to_stderr(next(f).rstrip())
+                    print_to_stderr(next(f).decode("utf-8", errors="ignore").rstrip())
                 print_to_stderr("##[endgroup]")
             for _ in range(min(n, num_lines)):
-                print_to_stderr(next(f).rstrip())
+                print_to_stderr(next(f).decode("utf-8", errors="ignore").rstrip())
             print_to_stderr(f"FINISHED PRINTING LOG FILE of {test} ({file_path})")
         else:
             print_to_stderr(f"Expand the folded group to see the log file of {test}")
             print_to_stderr(f"##[group]PRINTING LOG FILE of {test} ({file_path})")
-            print_to_stderr(f.read())
+            print_to_stderr(f.read().decode("utf-8", errors="ignore"))
             print_to_stderr("##[endgroup]")
             print_to_stderr(f"FINISHED PRINTING LOG FILE of {test} ({file_path})")
         print_to_stderr("")
 
 
-def get_pytest_args(options, stepcurrent_key, is_cpp_test=False):
+def get_pytest_args(
+    options, stepcurrent_key, is_cpp_test=False, is_distributed_test=False
+):
     if RERUN_DISABLED_TESTS:
+        # Distributed tests are too slow, so running them x50 will cause the jobs to timeout after
+        # 3+ hours. So, let's opt for less number of reruns. We need at least 150 instances of the
+        # test every 2 weeks to satisfy the Rockset query (15 x 14 = 210). The same logic applies
+        # to ASAN, which is also slow
+        count = 15 if is_distributed_test or TEST_WITH_ASAN else 50
         # When under rerun-disabled-tests mode, run the same tests multiple times to determine their
         # flakiness status. Default to 50 re-runs
-        rerun_options = ["--flake-finder", "--flake-runs=50"]
+        rerun_options = ["--flake-finder", f"--flake-runs={count}"]
     elif options.continue_through_error:
         # If continue through error, don't stop on first failure
         rerun_options = ["--reruns=2"]
@@ -956,9 +976,9 @@ def get_pytest_args(options, stepcurrent_key, is_cpp_test=False):
     ]
     if not is_cpp_test:
         # C++ tests need to be run with pytest directly, not via python
-        pytest_args.extend(
-            ["-p", "no:xdist", "--use-pytest", f"--sc={stepcurrent_key}"]
-        )
+        pytest_args.extend(["-p", "no:xdist", "--use-pytest"])
+        if not options.continue_through_error:
+            pytest_args.append(f"--sc={stepcurrent_key}")
     else:
         # Use pytext-dist to run C++ tests in parallel as running them sequentially using run_test
         # is much slower than running them directly
@@ -1255,8 +1275,8 @@ def exclude_tests(
 def must_serial(file: str) -> bool:
     return (
         os.getenv("PYTORCH_TEST_RUN_EVERYTHING_IN_SERIAL", "0") == "1"
-        or "distributed" in os.getenv("TEST_CONFIG", "")
-        or "distributed" in file
+        or DISTRIBUTED_TEST_PREFIX in os.getenv("TEST_CONFIG", "")
+        or DISTRIBUTED_TEST_PREFIX in file
         or file in CUSTOM_HANDLERS
         or file in RUN_PARALLEL_BLOCKLIST
         or file in CI_SERIAL_LIST
@@ -1414,7 +1434,11 @@ def get_selected_tests(options) -> List[ShardedTest]:
     # Do sharding
     test_file_times_config = test_file_times.get(test_config, {})
     shards = calculate_shards(
-        num_shards, selected_tests, test_file_times_config, must_serial=must_serial
+        num_shards,
+        selected_tests,
+        test_file_times_config,
+        must_serial=must_serial,
+        debug=TEST_WITH_ROCM,
     )
     _, tests_from_shard = shards[which_shard - 1]
     selected_tests = tests_from_shard
@@ -1497,7 +1521,11 @@ def run_tests(
 
     def parallel_test_completion_callback(err_message):
         test_failed = handle_error_messages(err_message)
-        if test_failed and not options.continue_through_error:
+        if (
+            test_failed
+            and not options.continue_through_error
+            and not RERUN_DISABLED_TESTS
+        ):
             pool.terminate()
 
     try:
@@ -1515,7 +1543,11 @@ def run_tests(
         pool.join()
         del os.environ["NUM_PARALLEL_PROCS"]
 
-        if not options.continue_through_error and len(failure_messages) != 0:
+        if (
+            not options.continue_through_error
+            and not RERUN_DISABLED_TESTS
+            and len(failure_messages) != 0
+        ):
             raise RuntimeError(
                 "\n".join(failure_messages)
                 + "\n\nTip: You can keep running tests even on failure by "
@@ -1530,7 +1562,11 @@ def run_tests(
                 options_clone.pytest = True
             err_message = run_test_module(test, test_directory, options_clone)
             test_failed = handle_error_messages(err_message)
-            if test_failed and not options.continue_through_error:
+            if (
+                test_failed
+                and not options.continue_through_error
+                and not RERUN_DISABLED_TESTS
+            ):
                 raise RuntimeError(err_message)
 
     finally:
@@ -1579,6 +1615,13 @@ def main():
     remaining_tests = selected_tests
     if IS_CI:
         (prioritized_tests, remaining_tests) = get_reordered_tests(selected_tests)
+        log_time_savings(
+            selected_tests,
+            prioritized_tests,
+            is_serial_test_fn=must_serial,
+            num_procs=NUM_PROCS,
+        )
+
         # downloading test cases configuration to local environment
         get_test_case_configs(dirpath=test_directory)
 
@@ -1617,7 +1660,10 @@ def main():
     if len(failure_messages) != 0:
         for err in failure_messages:
             print_to_stderr(err)
-        sys.exit(1)
+
+        # A disabled test is expected to fail, so there is no need to report a failure here
+        if not RERUN_DISABLED_TESTS:
+            sys.exit(1)
 
 
 if __name__ == "__main__":
